requirements.txt
torch
--upgrade torch
transformers
opencv-python
numpy
pandas
tqdm
Pillow
accelerate
openai
bitsandbytes